added interface update

+++ interfaces/jni/src/Const.cpp
@@ -6,6 +6,6 @@
 using namespace souffle;
 
 void Java_com_soufflelang_souffle_Const_init(JNIEnv* env, jobject obj, jlong num) {
-    AstNumberConstant* c = new AstNumberConstant((long)num);
+    AstNumberConstant* c = new AstNumberConstant((int32_t)num);
     setHandle(env, obj, c);
 }
